+++ plotnine/tests/test_stat_calculate_methods.py
@@ -2,9 +2,10 @@ from __future__ import absolute_import, division, print_function
 
 import six
 import pandas as pd
+import numpy as np
 import pytest
 
-from plotnine import ggplot, aes, stat_bin
+from plotnine import ggplot, aes, stat_bin, stat_density, xlim
 from plotnine.exceptions import PlotnineError
 
 
@@ -29,3 +30,15 @@ def test_stat_bin():
     gg = ggplot(aes(x='x', y='y'), df) + stat_bin()
     with pytest.raises(PlotnineError):
         gg.draw_test()
+
+
+def test_changing_xlim_in_stat_density():
+    n = 100
+    _xlim = (5, 10)
+    df = pd.DataFrame({'x': np.linspace(_xlim[0]-1, _xlim[1]+1, n)})
+    p = (ggplot(df, aes('x'))
+         + stat_density()
+         + xlim(*_xlim)
+         )
+    # No exceptions
+    p._build()