+++ doc/changelog.rst
@@ -14,12 +14,15 @@ v0.2.1
 
 - Fixed bug in :class:`~plotnine.stat_summary` where the you got
   an exception for some types of the `x` aesthetic values.
-  
+
 - Fixed bug where ``ggplot(data=df)`` resulted in an exception.
 
 - Fixed missing axis ticks and labels for :class:`~plotnine.facet_wrap`
   when the scales are allowed to vary (e.g `scales='free'`) between
   the panels.
+
+- Fixed bug in :class:`~plotnine.stat_density` where changing the
+  x limits lead to an exception (:issue: `22`)
 
 
 v0.2.0